refactored server JobTitle class to LcRest pattern #433
